Order repos

+++ Shamazon-BackEnd/Models/Products.cs
@@ -6,7 +6,7 @@
 
         public string ProductName { get; set; }
 
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
 
         public string ProductDescription { get; set; }
